# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_api.CreateEventRequest:
    properties:
      end_date:
        type: string
      event_type:
        type: string
      requires_attendance:
        type: boolean
      start_date:
        type: string
      title:
        type: string
    type: object
...
  internal_api.EventItem:
    properties:
      end_date:
        type: string
      event_type:
        type: string
      id:
        type: string
      requires_attendance:
        type: boolean
      start_date:
        type: string
      title:
        type: string
    type: object
...
  internal_api.UpdateEventRequest:
    properties:
      end_date:
        type: string
      event_type:
        type: string
      requires_attendance:
        type: boolean
      start_date:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Platform team
  description: Backend REST API for Somotracker educational dashboard
  license:
    name: MIT
  termsOfService: http://somotracker.local/terms
  title: Somotracker API
  version: 1.0.0
paths:
...
  /events:
    get:
      description: List upcoming school events for active school
      parameters:
      - description: Start date YYYY-MM-DD
        in: query
        name: from
        type: string
      - description: End date YYYY-MM-DD
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api.EventItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List events
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: Create a new school event
      parameters:
      - description: Event data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api.EventItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create event
      tags:
      - Events
  /events/{id}:
    delete:
      description: Delete a school event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete event
      tags:
      - Events
    patch:
      consumes:
      - application/json
      description: Update a school event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Event data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.EventItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update event
      tags:
      - Events
...
swagger: "2.0"